refactor: <useHook>, Main page 적용 및 파일 나누기
